feat: rename model, delete old
