+++ FRW/Modules/Login/ViewModel/KeyStoreLoginViewModel.swift
@@ -2,37 +2,37 @@
 //  KeyStoreLoginViewModel.swift
 //  FRW
 //
-//  Created by cat on 2024/8/19.
+//  Refactored version using LoginViewModelProtocol
+//  Created by cat on 2024/12/28.
 //
-
-import Foundation
-import WalletCore
-import Web3Core
 
 import Flow
 import FlowWalletKit
+import Foundation
 import SwiftUI
+import WalletCore
+import Web3Core
 
 // MARK: - KeyStoreLoginViewModel
 
-final class KeyStoreLoginViewModel_Old: ObservableObject {
-    // MARK: Internal
+final class KeyStoreLoginViewModel: ObservableObject, LoginViewModelProtocol {
+    // MARK: - LoginViewModelProtocol Required Properties
 
-    @Published
-    var json: String = ""
-    @Published
-    var password: String = ""
-    @Published
-    var wantedAddress: String = ""
+    typealias KeyType = FlowWalletKit.PrivateKey
 
-    @Published
-    var buttonState: VPrimaryButtonState = .disabled
+    @Published var wantedAddress: String = ""
+    @Published var buttonState: VPrimaryButtonState = .disabled
+    @Published var wallet: FlowWalletKit.Wallet? = nil
+    var cryptoKey: FlowWalletKit.PrivateKey?
+    var account: Flow.Account? = nil
 
+    // MARK: - Specific Properties
+
+    @Published var json: String = ""
+    @Published var password: String = ""
     var userName: String = ""
 
-    @Published
-    var wallet: FlowWalletKit.Wallet? = nil
-
+    // MARK: - UI Update Methods
 
     @MainActor
     func update(json _: String) {
@@ -46,157 +46,6 @@ final class KeyStoreLoginViewModel_Old: ObservableObject {
 
     func update(address _: String) {}
 
-    func onSumbit() {
-        UIApplication.shared.endEditing()
-        HUD.loading()
-        Task {
-            do {
-                privateKey = try PrivateKey.restore(
-                    json: json,
-                    password: password,
-                    storage: FlowWalletKit.PrivateKey.PKStorage
-                )
-                guard let privateKey else {
-                    HUD.error(title: "invalid_data".localized)
-                    return
-                }
-                wallet = FlowWalletKit.Wallet(type: .key(privateKey))
-
-                try await fetchAllAddresses()
-                HUD.dismissLoading()
-
-                if wantedAddress.isEmpty {
-                  guard let account = wallet?.flowAccounts?[currentNetwork]?.first else {
-                    return
-                  }
-                  selectedAccount(by: account)
-                } else {
-                    guard let keys = wallet?.flowAccounts?[currentNetwork] else {
-                        HUD.error(title: "not_find_address".localized)
-                        return
-                    }
-                    guard let account = keys.filter({ $0.address.hex == wantedAddress }).first
-                    else {
-                        HUD.error(title: "not_find_address".localized)
-                        return
-                    }
-                    selectedAccount(by: account)
-                }
-
-            } catch let error as FlowWalletKit.FWKError {
-                if error == FlowWalletKit.FWKError.invaildKeyStorePassword {
-                    HUD.error(title: "invalid_password".localized)
-                } else if error == FlowWalletKit.FWKError.invaildKeyStoreJSON {
-                    HUD.error(title: "invalid_json".localized)
-                }else {
-                    HUD.error(title: "invalid_data".localized)
-                }
-                HUD.dismissLoading()
-            } catch {
-                HUD.error(title: "invalid_data".localized)
-                HUD.dismissLoading()
-            }
-        }
-    }
-
-    // fetch all addresses of Public Key
-    func fetchAllAddresses() async throws {
-        do {
-            _ = try await wallet?.fetchAllNetworkAccounts()
-        } catch {
-            log.error("\(error.localizedDescription)")
-        }
-    }
-
-    func selectedAccount(by account: Flow.Account) {
-        self.account = account
-        checkPublicKey()
-    }
-
-    func createUserName(callback: @escaping (String) -> Void) {
-        let viewModel = ImportUserNameViewModel { name in
-            if !name.isEmpty {
-                callback(name)
-            }
-        }
-        Router.route(to: RouteMap.RestoreLogin.importUserName(viewModel))
-    }
-
-    func checkPublicKey() {
-        let keys = account?.keys.filter {
-                $0.publicKey.description == p256PublicKey || $0.publicKey
-                    .description == secp256PublicKey
-            }
-        guard let selectedKey = keys?.first,
-              let address = account?.address.hex, let privateKey = privateKey
-        else {
-            HUD.error(title: "not_find_address".localized)
-            log.error("[Import] keys of account not match the public:\(String(describing: p256PublicKey)) or \(String(describing: secp256PublicKey)) ")
-            return
-        }
-        guard selectedKey.weight >= 1000 else {
-            HUD.error(title: "account_key_weight_less".localized)
-            return
-        }
-        guard !selectedKey.revoked else {
-            HUD.error(title: "account_key_done_revoked_tips".localized)
-            return
-        }
-        Task {
-            HUD.loading()
-            do {
-                let publicKey = selectedKey.publicKey.description
-                let response: Network.EmptyResponse = try await Network
-                    .requestWithRawModel(FRWAPI.User.checkimport(publicKey))
-                if response.httpCode == 409 {
-                    try await UserManager.shared.importLogin(
-                        by: address,
-                        userName: "",
-                        flowKey: selectedKey,
-                        privateKey: privateKey
-                    )
-                } else if response.httpCode == 200 {
-                    createUserName { name in
-                        Task {
-                            try await UserManager.shared.importLogin(
-                                by: address,
-                                userName: name,
-                                flowKey: selectedKey,
-                                privateKey: privateKey,
-                                isImport: true
-                            )
-                            Router.popToRoot()
-                        }
-                    }
-                }
-                HUD.dismissLoading()
-            } catch {
-                if let code = error.moyaCode() {
-                    if code == 409 {
-                        do {
-                            try await UserManager.shared.importLogin(
-                                by: address,
-                                userName: "",
-                                flowKey: selectedKey,
-                                privateKey: privateKey
-                            )
-                            Router.popToRoot()
-                        } catch {
-                            log.error("[Import] login 409 :\(error)")
-                        }
-                    }
-                }
-                log.error("[Import] check public key own error:\(error)")
-                HUD.dismissLoading()
-            }
-        }
-    }
-
-    // MARK: Private
-
-    private var privateKey: FlowWalletKit.PrivateKey?
-    private var account: Flow.Account? = nil
-
     @MainActor
     private func update() {
         updateButtonState()
@@ -205,34 +54,86 @@ final class KeyStoreLoginViewModel_Old: ObservableObject {
     private func updateButtonState() {
         buttonState = (json.isEmpty || password.isEmpty) ? .disabled : .enabled
     }
-}
 
-extension KeyStoreLoginViewModel_Old {
-    private var p256PublicKey: String? {
-        privateKey?.publicKey(signAlgo: .ECDSA_P256)?.hexValue
+    // MARK: - LoginViewModelProtocol Required Methods
+
+    func onSubmit() {
+        UIApplication.shared.endEditing()
+        HUD.loading()
+
+        Task {
+            do {
+                // Restore private key from keystore JSON
+                cryptoKey = try PrivateKey.restore(
+                    json: json,
+                    password: password,
+                    storage: FlowWalletKit.PrivateKey.PKStorage
+                )
+
+                guard cryptoKey != nil else {
+                    HUD.error(title: "invalid_data".localized)
+                    HUD.dismissLoading()
+                    return
+                }
+
+                // Create wallet with restored key
+                wallet = FlowWalletKit.Wallet(type: .key(cryptoKey!))
+
+                // Fetch all addresses (common logic from protocol)
+                try await fetchAllAddresses()
+
+                HUD.dismissLoading()
+
+                // Select account (common logic from protocol)
+                selectAccountFromWallet()
+
+            } catch let error as FlowWalletKit.FWKError {
+                handleFWKError(error)
+                HUD.dismissLoading()
+            } catch {
+                HUD.error(title: "invalid_data".localized)
+                HUD.dismissLoading()
+            }
+        }
     }
 
-    private var secp256PublicKey: String? {
-        privateKey?.publicKey(signAlgo: .ECDSA_SECP256k1)?.hexValue
+    func getP256PublicKey() -> String? {
+        cryptoKey?.publicKey(signAlgo: .ECDSA_P256)?.hexValue
     }
-}
 
-// MARK: - Keystore
+    func getSecp256PublicKey() -> String? {
+        cryptoKey?.publicKey(signAlgo: .ECDSA_SECP256k1)?.hexValue
+    }
 
-struct Keystore: Codable {
-    var address: String?
-    var crypto: CryptoParamsV3
-    var id: String?
-    var version: Int
-}
+    func performLogin(
+        address: String,
+        userName: String,
+        flowKey: Flow.AccountKey,
+        isImport: Bool
+    ) async throws {
+        guard let privateKey = cryptoKey else {
+            throw LoginError.missingKey
+        }
 
-// MARK: - ImportAccountInfo
+        try await UserManager.shared.importLogin(
+            by: address,
+            userName: userName,
+            flowKey: flowKey,
+            privateKey: privateKey,
+            isImport: isImport
+        )
+    }
 
-struct ImportAccountInfo {
-    let address: String?
-    let weight: Int?
-    let keyId: Int?
-    let publicKey: String?
-    let signAlgo: Flow.SignatureAlgorithm
-    let hashAlgo: Flow.HashAlgorithm = .SHA2_256
+    // MARK: - Error Handling
+
+    private func handleFWKError(_ error: FlowWalletKit.FWKError) {
+        switch error {
+        case .invaildKeyStorePassword:
+            HUD.error(title: "invalid_password".localized)
+        case .invaildKeyStoreJSON:
+            HUD.error(title: "invalid_json".localized)
+        default:
+            HUD.error(title: "invalid_data".localized)
+        }
+    }
 }
