+++ FRW/Modules/Login/ViewModel/PrivateKeyLoginViewModel.swift
@@ -2,7 +2,8 @@
 //  PrivateKeyLoginViewModel.swift
 //  FRW
 //
-//  Created by cat on 2024/8/19.
+//  Refactored version using LoginViewModelProtocol
+//  Created by cat on 2024/12/28.
 //
 
 import CryptoKit
@@ -14,19 +15,29 @@ import WalletCore
 
 // MARK: - PrivateKeyLoginViewModel
 
-final class PrivateKeyLoginViewModel_Old: ObservableObject {
-    // MARK: Internal
+final class PrivateKeyLoginViewModel: ObservableObject, LoginViewModelProtocol {
+    // MARK: - LoginViewModelProtocol Required Properties
 
-    @Published
-    var key: String = ""
-    @Published
-    var wantedAddress: String = ""
-    @Published
-    var buttonState: VPrimaryButtonState = .disabled
+    typealias KeyType = FlowWalletKit.PrivateKey
 
+    @Published var wantedAddress: String = ""
+    @Published var buttonState: VPrimaryButtonState = .disabled
+    var wallet: FlowWalletKit.Wallet? = nil
+    var cryptoKey: FlowWalletKit.PrivateKey?
+    var account: Flow.Account? = nil
+
+    // MARK: - Specific Properties
+
+    @Published var key: String = ""
     var userName: String = ""
 
-    var wallet: FlowWalletKit.Wallet? = nil
+    // MARK: - Initialization
+
+    init() {
+        // Setup any initial state
+    }
+
+    // MARK: - UI Update Methods
 
     @MainActor
     func update(key _: String) {
@@ -38,167 +49,90 @@ final class PrivateKeyLoginViewModel_Old: ObservableObject {
         update()
     }
 
-    func onSumbit() {
-        UIApplication.shared.endEditing()
-        HUD.loading()
-        Task {
-            do {
-                let chainId = currentNetwork
-                guard let data = Data(hexString: key.stripHexPrefix()) else {
-                    HUD.dismissLoading()
-                    HUD.error(title: "invalid_data".localized)
-                    return
-                }
-
-                privateKey = try PrivateKey.restore(
-                    secret: data,
-                    storage: FlowWalletKit.PrivateKey.PKStorage
-                )
-                guard let privateKey = privateKey else {
-                    HUD.dismissLoading()
-                    HUD.error(title: "invalid_data".localized)
-                    return
-                }
-                wallet = FlowWalletKit.Wallet(type: .key(privateKey))
-
-                try await fetchAllAddresses()
-                HUD.dismissLoading()
-                if wantedAddress.isEmpty {
-                  guard let account = wallet?.flowAccounts?[currentNetwork]?.first else {
-                    return
-                  }
-                  selectedAccount(by: account)
-                } else {
-                    guard let keys = wallet?.flowAccounts?[currentNetwork] else {
-                        return
-                    }
-                    guard let account = keys.filter({ $0.address.hex == wantedAddress }).first
-                    else {
-                        HUD.error(title: "not_find_address".localized)
-                        return
-                    }
-                    selectedAccount(by: account)
-                }
-            } catch {
-                HUD.dismissLoading()
-            }
-        }
-    }
-
-    func fetchAllAddresses() async throws {
-        do {
-            _ = try await wallet?.fetchAllNetworkAccounts()
-        } catch {
-            log.error("\(error.localizedDescription)")
-        }
-    }
-
-    func selectedAccount(by account: Flow.Account) {
-        self.account = account
-        checkPublicKey()
-    }
-
-    func createUserName(callback: @escaping (String) -> Void) {
-        let viewModel = ImportUserNameViewModel { name in
-            if !name.isEmpty {
-                callback(name)
-            }
-        }
-        Router.route(to: RouteMap.RestoreLogin.importUserName(viewModel))
-    }
-
-    func checkPublicKey() {
-        let keys = account?.keys
-            .filter {
-                $0.publicKey.description == p256PublicKey || $0.publicKey
-                    .description == secp256PublicKey
-            }
-        guard let selectedKey = keys?.first,
-              let address = account?.address.hex, let privateKey = privateKey
-        else {
-            log.error("[Import] keys of account not match the public:\(String(describing: p256PublicKey)) or \(String(describing: secp256PublicKey)) ")
-            return
-        }
-        guard selectedKey.weight >= 1000 else {
-            HUD.error(title: "account_key_weight_less".localized)
-            return
-        }
-        guard !selectedKey.revoked else {
-            HUD.error(title: "account_key_done_revoked_tips".localized)
-            return
-        }
-        Task {
-            HUD.loading()
-            do {
-                let publicKey = selectedKey.publicKey.description
-                let response: Network.EmptyResponse = try await Network
-                    .requestWithRawModel(FRWAPI.User.checkimport(publicKey))
-                if response.httpCode == 409 {
-                    try await UserManager.shared.importLogin(
-                        by: address,
-                        userName: "",
-                        flowKey: selectedKey,
-                        privateKey: privateKey
-                    )
-                } else if response.httpCode == 200 {
-                    createUserName { _ in
-                        Task {
-                            try await UserManager.shared.importLogin(
-                                by: address,
-                                userName: self.userName,
-                                flowKey: selectedKey,
-                                privateKey: privateKey,
-                                isImport: true
-                            )
-                            Router.popToRoot()
-                        }
-                    }
-                }
-                HUD.dismissLoading()
-            } catch {
-                if let code = error.moyaCode() {
-                    if code == 409 {
-                        do {
-                            try await UserManager.shared.importLogin(
-                                by: address,
-                                userName: "",
-                                flowKey: selectedKey,
-                                privateKey: privateKey
-                            )
-                            Router.popToRoot()
-                        } catch {
-                            log.error("[Import] login 409 :\(error)")
-                        }
-                    }
-                }
-                log.error("[Import] check public key own error:\(error)")
-                HUD.dismissLoading()
-            }
-        }
-    }
-
-    // MARK: Private
-
-    private var privateKey: FlowWalletKit.PrivateKey?
-    private var account: Flow.Account? = nil
-
     @MainActor
     private func update() {
         updateButtonState()
     }
 
     private func updateButtonState() {
-        buttonState = (key.isEmpty) ? .disabled : .enabled
+        buttonState = key.isEmpty ? .disabled : .enabled
     }
 
+    // MARK: - LoginViewModelProtocol Required Methods
+
+    func onSubmit() {
+        UIApplication.shared.endEditing()
+        HUD.loading()
+
+        Task {
+            do {
+                // Restore private key from hex string
+                guard let data = Data(hexString: key.stripHexPrefix()) else {
+                    HUD.dismissLoading()
+                    HUD.error(title: "invalid_data".localized)
+                    return
+                }
+
+                cryptoKey = try PrivateKey.restore(
+                    secret: data,
+                    storage: FlowWalletKit.PrivateKey.PKStorage
+                )
+
+                guard cryptoKey != nil else {
+                    HUD.dismissLoading()
+                    HUD.error(title: "invalid_data".localized)
+                    return
+                }
+
+                // Create wallet with restored key
+                wallet = FlowWalletKit.Wallet(type: .key(cryptoKey!))
+
+                // Fetch all addresses (common logic from protocol)
+                try await fetchAllAddresses()
+
+                HUD.dismissLoading()
+
+                // Select account (common logic from protocol)
+                await selectAccountFromWallet()
+
+            } catch {
+                HUD.dismissLoading()
+                HUD.error(title: "invalid_data".localized)
+            }
+        }
+    }
+
+    func getP256PublicKey() -> String? {
+        cryptoKey?.publicKey(signAlgo: .ECDSA_P256)?.hexValue.format()
+    }
+
+    func getSecp256PublicKey() -> String? {
+        cryptoKey?.publicKey(signAlgo: .ECDSA_SECP256k1)?.hexValue.format()
+    }
+
+    func performLogin(
+        address: String,
+        userName: String,
+        flowKey: Flow.AccountKey,
+        isImport: Bool
+    ) async throws {
+        guard let privateKey = cryptoKey else {
+            throw LoginError.missingKey
+        }
+
+        try await UserManager.shared.importLogin(
+            by: address,
+            userName: userName,
+            flowKey: flowKey,
+            privateKey: privateKey,
+            isImport: isImport
+        )
+    }
 }
 
-extension PrivateKeyLoginViewModel_Old {
-    private var p256PublicKey: String? {
-        privateKey?.publicKey(signAlgo: .ECDSA_P256)?.hexValue.dropPrefix("04")
-    }
+// MARK: - LoginError
 
-    private var secp256PublicKey: String? {
-        privateKey?.publicKey(signAlgo: .ECDSA_SECP256k1)?.hexValue.dropPrefix("04")
-    }
+enum LoginError: Error {
+    case missingKey
+    case invalidData
 }
